Complete chapter 10 exercises

+++ test/integration/users_index_test.rb
@@ -5,6 +5,7 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
   def setup
     @admin     = users(:michael)
     @non_admin = users(:archer)
+    @not_active = users(:steveo)
   end
 
   test "index as admin including pagination and delete links" do
@@ -14,10 +15,12 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
     assert_select 'div.pagination'
     first_page_of_users = User.paginate(page: 1)
     first_page_of_users.each do |user|
-      assert_select 'a[href=?]', user_path(user), text: user.name
-      unless user == @admin
-        assert_select 'a[href=?]', user_path(user), text: 'delete',
-                                                    method: :delete
+      if user.activated?
+        assert_select 'a[href=?]', user_path(user), text: user.name
+        unless user == @admin
+          assert_select 'a[href=?]', user_path(user), text: 'delete',
+                                                      method: :delete
+        end
       end
     end
     assert_difference 'User.count', -1 do
@@ -29,5 +32,12 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
     log_in_as(@non_admin)
     get users_path
     assert_select 'a', text: 'delete', count: 0
+  end
+
+  test "show inactive user redirects to homepage" do
+    log_in_as(@non_admin)
+    get user_path(@not_active)
+    follow_redirect!
+    assert_template 'static_pages/home'
   end
 end
